requirements.txt
opencv-python>=4.5.0
Pillow>=8.0.0
numpy>=1.21.0
tensorflow>=2.8.0
scikit-image>=0.18.0
matplotlib>=3.3.0
requests>=2.25.0
click>=8.0.0
tqdm>=4.60.0